[v0.88][maintainability] Split cli/tests/pr_cmd_inline/finish/guardrails.rs by responsibility (Closes #1825)
