+++ adl/src/cli/tests/pr_cmd_inline/finish/guardrails/branch_and_gitignore.rs
@@ -1,0 +1,311 @@
+use super::*;
+
+#[test]
+fn real_pr_finish_rejects_main_and_reports_no_pr_when_only_local_bundle_sync_changes_exist() {
+    let _guard = env_lock();
+    let temp = unique_temp_dir("adl-pr-finish-errors");
+    let origin = temp.join("origin.git");
+    let repo = temp.join("repo");
+    fs::create_dir_all(&repo).expect("repo dir");
+    copy_bootstrap_support_files(&repo);
+    init_git_repo(&repo);
+    assert!(Command::new("git")
+        .args(["config", "user.name", "Test User"])
+        .current_dir(&repo)
+        .status()
+        .expect("git config")
+        .success());
+    assert!(Command::new("git")
+        .args(["config", "user.email", "test@example.com"])
+        .current_dir(&repo)
+        .status()
+        .expect("git config")
+        .success());
+    fs::write(repo.join(".gitignore"), ".adl/\n").expect("seed gitignore");
+    assert!(Command::new("git")
+        .args(["branch", "-M", "main"])
+        .current_dir(&repo)
+        .status()
+        .expect("git branch")
+        .success());
+    fs::create_dir_all(repo.join("adl/src")).expect("adl src");
+    fs::write(repo.join("adl/src/lib.rs"), "pub fn placeholder() {}\n").expect("write source");
+    let issue_ref =
+        IssueRef::new(1153, "v0.86".to_string(), "rust-finish-test".to_string()).expect("ref");
+    let bundle_dir = issue_ref.task_bundle_dir_path(&repo);
+    fs::create_dir_all(&bundle_dir).expect("bundle dir");
+    fs::write(issue_ref.task_bundle_input_path(&repo), "# input\n").expect("input");
+    write_completed_sor_fixture(
+        &issue_ref.task_bundle_output_path(&repo),
+        "codex/1153-rust-finish-test",
+    );
+    assert!(Command::new("git")
+        .args(["add", ".gitignore", "adl/src/lib.rs"])
+        .current_dir(&repo)
+        .status()
+        .expect("git add")
+        .success());
+    assert!(Command::new("git")
+        .args(["commit", "-q", "-m", "init on main"])
+        .current_dir(&repo)
+        .status()
+        .expect("git commit")
+        .success());
+    assert!(Command::new("git")
+        .args([
+            "init",
+            "--bare",
+            "-q",
+            path_str(&origin).expect("origin path"),
+        ])
+        .current_dir(&repo)
+        .status()
+        .expect("git init bare")
+        .success());
+    assert!(Command::new("git")
+        .args([
+            "remote",
+            "set-url",
+            "origin",
+            path_str(&origin).expect("origin path"),
+        ])
+        .current_dir(&repo)
+        .status()
+        .expect("git remote set-url")
+        .success());
+    assert!(Command::new("git")
+        .args(["push", "-q", "-u", "origin", "main"])
+        .current_dir(&repo)
+        .status()
+        .expect("git push")
+        .success());
+    let issue_ref =
+        IssueRef::new(1153, "v0.86".to_string(), "rust-finish-test".to_string()).expect("ref");
+    let bundle_dir = issue_ref.task_bundle_dir_path(&repo);
+    fs::create_dir_all(&bundle_dir).expect("bundle dir");
+    let stp = issue_ref.task_bundle_stp_path(&repo);
+    let input = issue_ref.task_bundle_input_path(&repo);
+    let output = issue_ref.task_bundle_output_path(&repo);
+    write_authored_issue_prompt(&repo, &issue_ref, "Example");
+    fs::copy(issue_ref.issue_prompt_path(&repo), &stp).expect("seed stp");
+    write_authored_sip(
+        &input,
+        &issue_ref,
+        "Example",
+        "codex/1153-rust-finish-test",
+        &issue_ref.issue_prompt_path(&repo),
+        &repo,
+    );
+    write_completed_sor_fixture(&output, "codex/1153-rust-finish-test");
+    let cards_root = resolve_cards_root(&repo, None);
+    let compat_output = card_output_path(&cards_root, 1153);
+    ensure_symlink(&compat_output, &output).expect("compat symlink");
+    assert!(Command::new("git")
+        .args(["add", "-A"])
+        .current_dir(&repo)
+        .status()
+        .expect("git add")
+        .success());
+    assert!(Command::new("git")
+        .args(["commit", "-q", "-m", "seed finish bundle"])
+        .current_dir(&repo)
+        .status()
+        .expect("git commit")
+        .success());
+    assert!(Command::new("git")
+        .args(["push", "-q", "origin", "main"])
+        .current_dir(&repo)
+        .status()
+        .expect("git push")
+        .success());
+
+    let prev_dir = env::current_dir().expect("cwd");
+    env::set_current_dir(&repo).expect("chdir");
+    let main_err = real_pr(&[
+        "finish".to_string(),
+        "1153".to_string(),
+        "--title".to_string(),
+        "Example".to_string(),
+        "--no-checks".to_string(),
+        "--no-open".to_string(),
+    ])
+    .expect_err("main should be rejected");
+    assert!(main_err.to_string().contains("refusing to run on main"));
+    env::set_current_dir(prev_dir).expect("restore cwd");
+
+    assert!(Command::new("git")
+        .args(["checkout", "-q", "-b", "codex/1153-rust-finish-test"])
+        .current_dir(&repo)
+        .status()
+        .expect("git checkout")
+        .success());
+    let mut touched_output = fs::read_to_string(&output).expect("read output");
+    touched_output.push('\n');
+    fs::write(&output, touched_output).expect("touch output");
+
+    let bin_dir = temp.join("bin");
+    fs::create_dir_all(&bin_dir).expect("bin dir");
+    let gh_path = bin_dir.join("gh");
+    write_executable(
+        &gh_path,
+        "#!/usr/bin/env bash\nset -euo pipefail\ncmd=\"$(printf '%s ' \"$@\")\"\nif printf '%s' \"$cmd\" | grep -q 'repo view'; then\n  printf 'danielbaustin/agent-design-language\\n'\n  exit 0\nfi\nif printf '%s' \"$cmd\" | grep -q 'pr list'; then\n  exit 0\nfi\nif printf '%s' \"$cmd\" | grep -q 'pr view'; then\n  if printf '%s' \"$cmd\" | grep -q 'closingIssuesReferences'; then\n    printf '1153\\n'\n  else\n    printf 'Closes #1153\\n'\n  fi\n  exit 0\nfi\nexit 1\n",
+    );
+    let old_path = env::var("PATH").unwrap_or_default();
+    let prev_dir = env::current_dir().expect("cwd");
+    unsafe {
+        env::set_var("PATH", format!("{}:{}", bin_dir.display(), old_path));
+    }
+    env::set_current_dir(&repo).expect("chdir");
+    let bundle_sync_err = real_pr(&[
+        "finish".to_string(),
+        "1153".to_string(),
+        "--title".to_string(),
+        "Example".to_string(),
+        "--input".to_string(),
+        path_relative_to_repo(&repo, &input),
+        "--output".to_string(),
+        path_relative_to_repo(&repo, &output),
+        "--no-checks".to_string(),
+        "--no-open".to_string(),
+    ])
+    .expect_err("expected downstream gh fixture failure after bundle sync work is detected");
+    env::set_current_dir(prev_dir).expect("restore cwd");
+    unsafe {
+        env::set_var("PATH", old_path);
+    }
+    assert!(bundle_sync_err.to_string().contains("Nothing to PR."));
+}
+
+#[test]
+fn real_pr_finish_rejects_staged_gitignore_changes_without_allow_flag() {
+    let _guard = env_lock();
+    let temp = unique_temp_dir("adl-pr-finish-gitignore-guard");
+    let origin = temp.join("origin.git");
+    let repo = temp.join("repo");
+    fs::create_dir_all(&repo).expect("repo dir");
+    copy_bootstrap_support_files(&repo);
+    init_git_repo(&repo);
+    assert!(Command::new("git")
+        .args(["config", "user.name", "Test User"])
+        .current_dir(&repo)
+        .status()
+        .expect("git config")
+        .success());
+    assert!(Command::new("git")
+        .args(["config", "user.email", "test@example.com"])
+        .current_dir(&repo)
+        .status()
+        .expect("git config")
+        .success());
+    fs::write(repo.join(".gitignore"), ".adl/\n").expect("seed gitignore");
+    assert!(Command::new("git")
+        .args(["branch", "-M", "main"])
+        .current_dir(&repo)
+        .status()
+        .expect("git branch")
+        .success());
+    fs::create_dir_all(repo.join("adl/src")).expect("adl src");
+    fs::write(repo.join("adl/src/lib.rs"), "pub fn placeholder() {}\n").expect("write source");
+    let issue_ref =
+        IssueRef::new(1153, "v0.86".to_string(), "rust-finish-test".to_string()).expect("ref");
+    let bundle_dir = issue_ref.task_bundle_dir_path(&repo);
+    fs::create_dir_all(&bundle_dir).expect("bundle dir");
+    write_authored_issue_prompt(&repo, &issue_ref, "Example");
+    fs::copy(
+        issue_ref.issue_prompt_path(&repo),
+        issue_ref.task_bundle_stp_path(&repo),
+    )
+    .expect("seed stp");
+    write_authored_sip(
+        &issue_ref.task_bundle_input_path(&repo),
+        &issue_ref,
+        "Example",
+        "codex/1153-rust-finish-test",
+        &issue_ref.issue_prompt_path(&repo),
+        &repo,
+    );
+    write_completed_sor_fixture(
+        &issue_ref.task_bundle_output_path(&repo),
+        "codex/1153-rust-finish-test",
+    );
+    let cards_root = resolve_cards_root(&repo, None);
+    let compat_output = card_output_path(&cards_root, 1153);
+    ensure_symlink(&compat_output, &issue_ref.task_bundle_output_path(&repo))
+        .expect("compat symlink");
+    assert!(Command::new("git")
+        .args(["add", ".gitignore", "adl/src/lib.rs"])
+        .current_dir(&repo)
+        .status()
+        .expect("git add")
+        .success());
+    assert!(Command::new("git")
+        .args(["commit", "-q", "-m", "init"])
+        .current_dir(&repo)
+        .status()
+        .expect("git commit")
+        .success());
+    assert!(Command::new("git")
+        .args([
+            "init",
+            "--bare",
+            "-q",
+            path_str(&origin).expect("origin path"),
+        ])
+        .current_dir(&repo)
+        .status()
+        .expect("git init bare")
+        .success());
+    assert!(Command::new("git")
+        .args([
+            "remote",
+            "set-url",
+            "origin",
+            path_str(&origin).expect("origin path"),
+        ])
+        .current_dir(&repo)
+        .status()
+        .expect("git remote set-url")
+        .success());
+    assert!(Command::new("git")
+        .args(["push", "-q", "-u", "origin", "main"])
+        .current_dir(&repo)
+        .status()
+        .expect("git push")
+        .success());
+    assert!(Command::new("git")
+        .args(["checkout", "-q", "-b", "codex/1153-rust-finish-test"])
+        .current_dir(&repo)
+        .status()
+        .expect("git checkout")
+        .success());
+    fs::write(repo.join(".gitignore"), ".adl/\ntarget\n").expect("write gitignore");
+    assert!(Command::new("git")
+        .args(["add", ".gitignore"])
+        .current_dir(&repo)
+        .status()
+        .expect("git add gitignore")
+        .success());
+
+    let prev_dir = env::current_dir().expect("cwd");
+    env::set_current_dir(&repo).expect("chdir");
+    let err = real_pr(&[
+        "finish".to_string(),
+        "1153".to_string(),
+        "--title".to_string(),
+        "Example".to_string(),
+        "--input".to_string(),
+        path_relative_to_repo(&repo, &issue_ref.task_bundle_input_path(&repo)),
+        "--output".to_string(),
+        path_relative_to_repo(&repo, &issue_ref.task_bundle_output_path(&repo)),
+        "--no-checks".to_string(),
+        "--no-open".to_string(),
+    ])
+    .expect_err("gitignore guard should block finish");
+    env::set_current_dir(prev_dir).expect("restore cwd");
+    assert!(err
+        .to_string()
+        .contains("staged .gitignore or adl/.gitignore changes detected"));
+    assert!(err
+        .to_string()
+        .contains("Canonical .adl issue bundles are local-only and must not be staged"));
+}
